test: Move e2e tests into integration tests.
