+++ tests/serverpod_test_server/test_integration/database_operations/crud/delete_test.dart
@@ -8,6 +8,45 @@ void main() async {
 
   tearDown(() async {
     await SimpleData.db.deleteWhere(session, where: (_) => Constant.bool(true));
+    await UniqueData.db.deleteWhere(session, where: (_) => Constant.bool(true));
+  });
+
+  group('Given an empty database', () {
+    test(
+        'when trying to delete a row that doesn\'t exist then an error is thrown.',
+        () {
+      expect(
+        SimpleData.db.deleteRow(
+          session,
+          SimpleData(id: 1, num: 1),
+        ),
+        throwsA(isA<DatabaseDeleteRowException>()),
+      );
+    });
+
+    test(
+        'when trying to batch delete a row that doesn\'t exist then an error is thrown.',
+        () {
+      expect(
+        SimpleData.db.delete(
+          session,
+          [SimpleData(id: 1, num: 1)],
+        ),
+        throwsA(isA<DatabaseException>()),
+      );
+    },
+        skip:
+            'This call is not atomic but should be, reenable this test when bug is fixed.');
+
+    test('when trying to delete where (all) then an empty list is returned.',
+        () async {
+      var result = await SimpleData.db.deleteWhere(
+        session,
+        where: (t) => Constant.bool(true),
+      );
+
+      expect(result, []);
+    });
   });
 
   group('Given a list of entries', () {
@@ -86,103 +125,165 @@ void main() async {
     });
   });
 
-  test(
-      'Given two entries in the database when batch deleting the rows then the deleted ids are returned.',
-      () async {
-    var data = <UniqueData>[
-      UniqueData(number: 1, email: 'info@serverpod.dev'),
-      UniqueData(number: 2, email: 'dev@serverpod.dev'),
-    ];
+  group(' ', () {
+    tearDown(() async {
+      await RelatedUniqueData.db.deleteWhere(
+        session,
+        where: (_) => Constant.bool(true),
+      );
+      await SimpleData.db.deleteWhere(
+        session,
+        where: (_) => Constant.bool(true),
+      );
+      await UniqueData.db.deleteWhere(
+        session,
+        where: (_) => Constant.bool(true),
+      );
+    });
 
-    var inserted = await UniqueData.db.insert(session, data);
+    test(
+        'Given an inserted object when deleting that row then the id of the row is returned.',
+        () async {
+      var simpleData = SimpleData(num: 1);
+      var inserted = await SimpleData.db.insertRow(session, simpleData);
 
-    var deletedIds = await UniqueData.db.delete(session, data);
+      var deleted = await await SimpleData.db.deleteRow(
+        session,
+        inserted,
+      );
+      expect(deleted.id!, inserted.id);
+    });
 
-    expect(deletedIds.first.id!, inserted.first.id);
-    expect(deletedIds.last.id!, inserted.last.id);
-  });
+    test(
+        'Given an inserted object when deleting that row then it cannot be retrieved from the db.',
+        () async {
+      var simpleData = SimpleData(num: 1);
+      var inserted = await SimpleData.db.insertRow(session, simpleData);
+      var deleted = await SimpleData.db.deleteRow(session, inserted);
 
-  test(
-      'Given two entries in the database when batch deleting the rows then the rows are deleted from the database.',
-      () async {
-    var data = <UniqueData>[
-      UniqueData(number: 1, email: 'info@serverpod.dev'),
-      UniqueData(number: 2, email: 'dev@serverpod.dev'),
-    ];
+      var retrieved = await SimpleData.db.findById(session, deleted.id!);
 
-    var inserted = await UniqueData.db.insert(session, data);
+      expect(retrieved, isNull);
+    });
 
-    await UniqueData.db.delete(session, data);
+    test(
+        'Given two inserted objects when deleting all then the ids of the rows are returned.',
+        () async {
+      var simpleData1 = SimpleData(num: 1);
+      var simpleData2 = SimpleData(num: 2);
+      var inserted1 = await SimpleData.db.insertRow(session, simpleData1);
+      var inserted2 = await SimpleData.db.insertRow(session, simpleData2);
 
-    var first = await UniqueData.db.findById(session, inserted.first.id!);
-    var last = await UniqueData.db.findById(session, inserted.last.id!);
+      var deleted = await SimpleData.db.deleteWhere(
+        where: (t) => Constant.bool(true),
+        session,
+      );
 
-    expect(first, isNull);
-    expect(last, isNull);
-  });
+      var deletedIds = deleted.map((e) => e.id);
 
-  test(
-      'Given two entries in the database when batch deleting fails no rows are deleted from the database.',
-      () async {
-    var data = <UniqueData>[
-      UniqueData(number: 1, email: 'info@serverpod.dev'),
-      UniqueData(number: 2, email: 'dev@serverpod.dev'),
-    ];
+      expect(deletedIds, hasLength(2));
+      expect(deletedIds, contains(inserted1.id));
+      expect(deletedIds, contains(inserted2.id));
+    });
 
-    var inserted = await UniqueData.db.insert(session, data);
+    test(
+        'Given two entries in the database when batch deleting the rows then the deleted ids are returned.',
+        () async {
+      var data = <UniqueData>[
+        UniqueData(number: 1, email: 'info@serverpod.dev'),
+        UniqueData(number: 2, email: 'dev@serverpod.dev'),
+      ];
 
-    var relationalData = RelatedUniqueData(
-      number: 1,
-      uniqueDataId: inserted.last.id!,
-    );
+      var inserted = await UniqueData.db.insert(session, data);
 
-    // This restricts the delete of the second entry
-    RelatedUniqueData.db.insertRow(session, relationalData);
+      var deletedIds = await UniqueData.db.delete(session, inserted);
 
-    expect(
-      UniqueData.db.delete(session, inserted),
-      throwsA(isA<ArgumentError>()),
-    );
+      expect(deletedIds.first.id!, inserted.first.id);
+      expect(deletedIds.last.id!, inserted.last.id);
+    });
 
-    var first = await UniqueData.db.findById(session, inserted.first.id!);
-    var last = await UniqueData.db.findById(session, inserted.last.id!);
+    test(
+        'Given two entries in the database when batch deleting the rows then the rows are deleted from the database.',
+        () async {
+      var data = <UniqueData>[
+        UniqueData(number: 1, email: 'info@serverpod.dev'),
+        UniqueData(number: 2, email: 'dev@serverpod.dev'),
+      ];
 
-    expect(first, isNotNull);
-    expect(last, isNotNull);
-  });
+      var inserted = await UniqueData.db.insert(session, data);
 
-  test(
-      'Given two entries in the database when batch deleting one the other entry is still in the database.',
-      () async {
-    var data = <UniqueData>[
-      UniqueData(number: 1, email: 'info@serverpod.dev'),
-      UniqueData(number: 2, email: 'dev@serverpod.dev'),
-    ];
+      await UniqueData.db.delete(session, inserted);
 
-    var inserted = await UniqueData.db.insert(session, data);
+      var first = await UniqueData.db.findById(session, inserted.first.id!);
+      var last = await UniqueData.db.findById(session, inserted.last.id!);
 
-    await UniqueData.db.delete(session, [inserted.first]);
+      expect(first, isNull);
+      expect(last, isNull);
+    });
 
-    var first = await UniqueData.db.findById(session, inserted.first.id!);
-    var last = await UniqueData.db.findById(session, inserted.last.id!);
+    test(
+        'Given two entries in the database when batch deleting fails no rows are deleted from the database.',
+        () async {
+      var data = <UniqueData>[
+        UniqueData(number: 1, email: 'info@serverpod.dev'),
+        UniqueData(number: 2, email: 'dev@serverpod.dev'),
+      ];
 
-    expect(first, isNull);
-    expect(last, isNotNull);
-  });
+      var inserted = await UniqueData.db.insert(session, data);
 
-  test(
-      'Given two entries in the database when batch deleting one only that id is returned.',
-      () async {
-    var data = <UniqueData>[
-      UniqueData(number: 1, email: 'info@serverpod.dev'),
-      UniqueData(number: 2, email: 'dev@serverpod.dev'),
-    ];
+      var relationalData = RelatedUniqueData(
+        number: 1,
+        uniqueDataId: inserted.last.id!,
+      );
 
-    var inserted = await UniqueData.db.insert(session, data);
+      // This restricts the delete of the second entry
+      RelatedUniqueData.db.insertRow(session, relationalData);
 
-    var deleted = await UniqueData.db.delete(session, [inserted.first]);
+      expect(
+        UniqueData.db.delete(session, inserted),
+        throwsA(isA<DatabaseException>()),
+      );
 
-    expect(deleted, hasLength(1));
-    expect(deleted.first.id!, inserted.first.id);
+      var first = await UniqueData.db.findById(session, inserted.first.id!);
+      var last = await UniqueData.db.findById(session, inserted.last.id!);
+
+      expect(first, isNotNull);
+      expect(last, isNotNull);
+    });
+
+    test(
+        'Given two entries in the database when batch deleting one the other entry is still in the database.',
+        () async {
+      var data = <UniqueData>[
+        UniqueData(number: 1, email: 'info@serverpod.dev'),
+        UniqueData(number: 2, email: 'dev@serverpod.dev'),
+      ];
+
+      var inserted = await UniqueData.db.insert(session, data);
+
+      await UniqueData.db.delete(session, [inserted.first]);
+
+      var first = await UniqueData.db.findById(session, inserted.first.id!);
+      var last = await UniqueData.db.findById(session, inserted.last.id!);
+
+      expect(first, isNull);
+      expect(last, isNotNull);
+    });
+
+    test(
+        'Given two entries in the database when batch deleting one only that id is returned.',
+        () async {
+      var data = <UniqueData>[
+        UniqueData(number: 1, email: 'info@serverpod.dev'),
+        UniqueData(number: 2, email: 'dev@serverpod.dev'),
+      ];
+
+      var inserted = await UniqueData.db.insert(session, data);
+
+      var deleted = await UniqueData.db.delete(session, [inserted.first]);
+
+      expect(deleted, hasLength(1));
+      expect(deleted.first.id!, inserted.first.id);
+    });
   });
 }
