+++ tests/serverpod_test_server/test_integration/database_operations/crud/update_test.dart
@@ -1,11 +1,8 @@
 import 'dart:typed_data';
-
-import 'package:serverpod/database.dart';
+import 'package:serverpod/serverpod.dart';
 import 'package:serverpod_test_server/src/generated/protocol.dart';
 import 'package:serverpod_test_server/test_util/test_serverpod.dart';
 import 'package:test/test.dart';
-import 'package:uuid/uuid.dart';
-import 'package:uuid/uuid_value.dart';
 
 void main() async {
   var session = await IntegrationTestServer().session();
@@ -100,7 +97,7 @@ void main() async {
   });
 
   test(
-      'Given an model when updatingRow with a specific column only that column and no other data is updated.',
+      'Given an model when batch updatingRow with a specific column only that column and no other data is updated.',
       () async {
     var expectedEmail = 'info@serverpod.dev';
     var expectedNumber = 5;
@@ -127,7 +124,8 @@ void main() async {
     expect(updated.email, expectedEmail);
   });
 
-  test('Given an model without an id when updatingRow then an error is thrown.',
+  test(
+      'Given an model without an id when batch updatingRow then an error is thrown.',
       () async {
     expect(
       UniqueData.db.updateRow(
@@ -158,10 +156,299 @@ void main() async {
   });
 
   group('Given a typed entry in the database', () {
+    var data = Types(
+      anInt: 1,
+      aBool: true,
+      aDouble: 1.0,
+      aString: 'string',
+      aDateTime: DateTime.now(),
+      aByteData: ByteData.view(Uint8List(8).buffer),
+      aDuration: Duration(milliseconds: 1000),
+      aUuid: UuidValue.fromString(Uuid().v4()),
+      anEnum: TestEnum.one,
+    );
+
+    late Types type;
+
+    setUp(() async {
+      type = await Types.db.insertRow(session, data);
+    });
+
     tearDown(() async {
       await Types.db.deleteWhere(session, where: (t) => Constant.bool(true));
     });
 
+    test(
+        'when updating anInt to null then the database is updated with null value.',
+        () async {
+      var value = Types(
+        id: type.id,
+        anInt: null,
+      );
+
+      var updated = await Types.db.updateRow(session, value);
+
+      expect(updated.anInt, isNull);
+    });
+
+    test(
+        'when updating aBool to null then the database is updated with null value.',
+        () async {
+      var value = Types(
+        id: type.id,
+        aBool: null,
+      );
+
+      var updated = await Types.db.updateRow(session, value);
+
+      expect(updated.aBool, isNull);
+    });
+
+    test(
+        'when updating aDouble to null then the database is updated with null value.',
+        () async {
+      var value = Types(
+        id: type.id,
+        aDouble: null,
+      );
+
+      var updated = await Types.db.updateRow(session, value);
+
+      expect(updated.aDouble, isNull);
+    });
+
+    test(
+        'when updating aString to null then the database is updated with null value.',
+        () async {
+      var value = Types(
+        id: type.id,
+        aString: null,
+      );
+
+      var updated = await Types.db.updateRow(session, value);
+
+      expect(updated.aString, isNull);
+    });
+
+    test(
+        'when updating aDateTime to null then the database is updated with null value.',
+        () async {
+      var value = Types(
+        id: type.id,
+        aDateTime: null,
+      );
+
+      var updated = await Types.db.updateRow(session, value);
+
+      expect(updated.aDateTime, isNull);
+    });
+
+    test(
+        'when updating aByteData to null then the database is updated with null value.',
+        () async {
+      var value = Types(
+        id: type.id,
+        aByteData: null,
+      );
+
+      var updated = await Types.db.updateRow(session, value);
+
+      expect(updated.aByteData, isNull);
+    });
+
+    test(
+        'when updating aDuration to null then the database is updated with null value.',
+        () async {
+      var value = Types(
+        id: type.id,
+        aDuration: null,
+      );
+
+      var updated = await Types.db.updateRow(session, value);
+
+      expect(updated.aDuration, isNull);
+    });
+
+    test(
+        'when updating aUuid to null then the database is updated with null value.',
+        () async {
+      var value = Types(
+        id: type.id,
+        aUuid: null,
+      );
+
+      var updated = await Types.db.updateRow(session, value);
+
+      expect(updated.aUuid, isNull);
+    });
+
+    test(
+        'when updating anEnum to null then the database is updated with null value.',
+        () async {
+      var value = Types(
+        id: type.id,
+        anEnum: null,
+      );
+
+      var updated = await Types.db.updateRow(session, value);
+
+      expect(updated.anEnum, isNull);
+    });
+  });
+
+  group('Given a typed entry in the database', () {
+    var data = Types(
+      anInt: null,
+      aBool: null,
+      aDouble: null,
+      aString: null,
+      aDateTime: null,
+      aByteData: null,
+      aDuration: null,
+      aUuid: null,
+      anEnum: null,
+    );
+
+    late Types type;
+
+    setUp(() async {
+      type = await Types.db.insertRow(session, data);
+    });
+
+    tearDown(() async {
+      await Types.db.deleteWhere(session, where: (t) => Constant.bool(true));
+    });
+
+    test('when updating anInt to 1 then the database is updated with value 1.',
+        () async {
+      var value = Types(
+        id: type.id,
+        anInt: 1,
+      );
+
+      var updated = await Types.db.updateRow(session, value);
+
+      expect(updated.anInt, equals(1));
+    });
+
+    test(
+        'when updating aBool to true then the database is updated with value true.',
+        () async {
+      var value = Types(
+        id: type.id,
+        aBool: true,
+      );
+
+      var updated = await Types.db.updateRow(session, value);
+
+      expect(updated.aBool, equals(true));
+    });
+
+    test(
+        'when updating aDouble to 1.0 then the database is updated with value 1.0.',
+        () async {
+      var value = Types(
+        id: type.id,
+        aDouble: 1.0,
+      );
+
+      var updated = await Types.db.updateRow(session, value);
+
+      expect(updated.aDouble, equals(1.0));
+    });
+
+    test(
+        'when updating aString to "string" then the database is updated with value "string".',
+        () async {
+      var value = Types(
+        id: type.id,
+        aString: 'string',
+      );
+
+      var updated = await Types.db.updateRow(session, value);
+
+      expect(updated.aString, equals('string'));
+    });
+
+    test(
+        'when updating aDateTime to a real value then the database is updated with the real value.',
+        () async {
+      var now = DateTime.now().toUtc();
+      var value = Types(
+        id: type.id,
+        aDateTime: now,
+      );
+
+      var updated = await Types.db.updateRow(session, value);
+
+      expect(updated.aDateTime, equals(now));
+    });
+
+    test(
+        'when updating aByteData to a real value then the database is updated with the real value.',
+        () async {
+      var byteData = ByteData.view(Uint8List(8).buffer);
+      var value = Types(
+        id: type.id,
+        aByteData: byteData,
+      );
+
+      var updated = await Types.db.updateRow(session, value);
+
+      var byteDataResult = updated.aByteData;
+
+      expect(
+        byteDataResult?.buffer.asUint8List(
+          byteDataResult.offsetInBytes,
+          byteDataResult.lengthInBytes,
+        ),
+        equals(byteData.buffer.asUint8List()),
+      );
+    });
+
+    test(
+        'when updating aDuration to a real value then the database is updated with the real value.',
+        () async {
+      var duration = Duration(milliseconds: 1000);
+      var value = Types(
+        id: type.id,
+        aDuration: duration,
+      );
+
+      var updated = await Types.db.updateRow(session, value);
+
+      expect(updated.aDuration, equals(duration));
+    });
+
+    test(
+        'when updating aUuid to a real value then the database is updated with the real value.',
+        () async {
+      var uuidValue = UuidValue.fromString(Uuid().v4());
+      var value = Types(
+        id: type.id,
+        aUuid: uuidValue,
+      );
+
+      var updated = await Types.db.updateRow(session, value);
+
+      expect(updated.aUuid, equals(uuidValue));
+    });
+
+    test(
+        'when updating anEnum to TestEnum.one then the database is updated with value TestEnum.one.',
+        () async {
+      var value = Types(
+        id: type.id,
+        anEnum: TestEnum.one,
+      );
+
+      var updated = await Types.db.updateRow(session, value);
+
+      expect(updated.anEnum, equals(TestEnum.one));
+    });
+  });
+
+  group('Given a typed entry in the database', () {
     var data = <Types>[
       Types(
         anInt: 1,
@@ -183,8 +470,12 @@ void main() async {
       type = inserted.first;
     });
 
+    tearDown(() async {
+      await Types.db.deleteWhere(session, where: (t) => Constant.bool(true));
+    });
+
     test(
-        'when updating anInt to null then the database is updated with null value.',
+        'when batch updating anInt to null then the database is updated with null value.',
         () async {
       var toUpdate = <Types>[
         Types(
@@ -199,7 +490,7 @@ void main() async {
     });
 
     test(
-        'when updating aBool to null then the database is updated with null value.',
+        'when batch updating aBool to null then the database is updated with null value.',
         () async {
       var toUpdate = <Types>[
         Types(
@@ -214,7 +505,7 @@ void main() async {
     });
 
     test(
-        'when updating aDouble to null then the database is updated with null value.',
+        'when batch updating aDouble to null then the database is updated with null value.',
         () async {
       var toUpdate = <Types>[
         Types(
@@ -229,7 +520,7 @@ void main() async {
     });
 
     test(
-        'when updating aString to null then the database is updated with null value.',
+        'when batch updating aString to null then the database is updated with null value.',
         () async {
       var toUpdate = <Types>[
         Types(
@@ -244,7 +535,7 @@ void main() async {
     });
 
     test(
-        'when updating aDateTime to null then the database is updated with null value.',
+        'when batch updating aDateTime to null then the database is updated with null value.',
         () async {
       var toUpdate = <Types>[
         Types(
@@ -259,7 +550,7 @@ void main() async {
     });
 
     test(
-        'when updating aByteData to null then the database is updated with null value.',
+        'when batch updating aByteData to null then the database is updated with null value.',
         () async {
       var toUpdate = <Types>[
         Types(
@@ -274,7 +565,7 @@ void main() async {
     });
 
     test(
-        'when updating aDuration to null then the database is updated with null value.',
+        'when batch updating aDuration to null then the database is updated with null value.',
         () async {
       var toUpdate = <Types>[
         Types(
@@ -289,7 +580,7 @@ void main() async {
     });
 
     test(
-        'when updating aUuid to null then the database is updated with null value.',
+        'when batch updating aUuid to null then the database is updated with null value.',
         () async {
       var toUpdate = <Types>[
         Types(
@@ -304,7 +595,7 @@ void main() async {
     });
 
     test(
-        'when updating anEnum to null then the database is updated with null value.',
+        'when batch updating anEnum to null then the database is updated with null value.',
         () async {
       var toUpdate = <Types>[
         Types(
@@ -345,7 +636,8 @@ void main() async {
       await Types.db.deleteWhere(session, where: (t) => Constant.bool(true));
     });
 
-    test('when updating anInt to 1 then the database is updated with value 1.',
+    test(
+        'when batch updating anInt to 1 then the database is updated with value 1.',
         () async {
       var toUpdate = <Types>[
         Types(
@@ -360,7 +652,7 @@ void main() async {
     });
 
     test(
-        'when updating aBool to true then the database is updated with value true.',
+        'when batch updating aBool to true then the database is updated with value true.',
         () async {
       var toUpdate = <Types>[
         Types(
@@ -375,7 +667,7 @@ void main() async {
     });
 
     test(
-        'when updating aDouble to 1.0 then the database is updated with value 1.0.',
+        'when batch updating aDouble to 1.0 then the database is updated with value 1.0.',
         () async {
       var toUpdate = <Types>[
         Types(
@@ -390,7 +682,7 @@ void main() async {
     });
 
     test(
-        'when updating aString to "string" then the database is updated with value "string".',
+        'when batch updating aString to "string" then the database is updated with value "string".',
         () async {
       var toUpdate = <Types>[
         Types(
@@ -405,7 +697,7 @@ void main() async {
     });
 
     test(
-        'when updating aDateTime to a real value then the database is updated with the real value.',
+        'when batch updating aDateTime to a real value then the database is updated with the real value.',
         () async {
       var now = DateTime.now().toUtc();
       var toUpdate = <Types>[
@@ -421,7 +713,7 @@ void main() async {
     });
 
     test(
-        'when updating aByteData to a real value then the database is updated with the real value.',
+        'when batch updating aByteData to a real value then the database is updated with the real value.',
         () async {
       var byteData = ByteData.view(Uint8List(8).buffer);
       var toUpdate = <Types>[
@@ -433,14 +725,19 @@ void main() async {
 
       var updated = await Types.db.update(session, toUpdate);
 
+      var byteDataResult = updated.first.aByteData;
+
       expect(
-        updated.first.aByteData?.buffer.asUint8List(),
+        byteDataResult?.buffer.asUint8List(
+          byteDataResult.offsetInBytes,
+          byteDataResult.lengthInBytes,
+        ),
         equals(byteData.buffer.asUint8List()),
       );
     });
 
     test(
-        'when updating aDuration to a real value then the database is updated with the real value.',
+        'when batch updating aDuration to a real value then the database is updated with the real value.',
         () async {
       var duration = Duration(milliseconds: 1000);
       var toUpdate = <Types>[
@@ -456,7 +753,7 @@ void main() async {
     });
 
     test(
-        'when updating aUuid to a real value then the database is updated with the real value.',
+        'when batch updating aUuid to a real value then the database is updated with the real value.',
         () async {
       var uuidValue = UuidValue.fromString(Uuid().v4());
       var toUpdate = <Types>[
@@ -472,7 +769,7 @@ void main() async {
     });
 
     test(
-        'when updating anEnum to TestEnum.one then the database is updated with value TestEnum.one.',
+        'when batch updating anEnum to TestEnum.one then the database is updated with value TestEnum.one.',
         () async {
       var toUpdate = <Types>[
         Types(
